feat(objects): add compiler.Objects submodule
