+++ storyscript/compiler/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
 from .compiler import Compiler
+from .objects import Objects
 
-__all__ = ['Compiler']
+__all__ = ['Compiler', 'Objects']
