+++ tests/unittests/compiler/conftest.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+from pytest import fixture
+
+
+@fixture
+def tree(magic):
+    return magic()
